+++ normalized-cache-sqlite/src/commonMain/kotlin/com/apollographql/cache/normalized/sql/SqlNormalizedCacheFactory.kt
@@ -8,9 +8,10 @@ import com.apollographql.cache.normalized.sql.internal.RecordDatabase
 /**
  * Creates a new [NormalizedCacheFactory] that uses a persistent cache based on Sqlite
  *
- * @param name: the name of the database or null for an in-memory database
+ * @param name The name of the database or null for an in-memory database
  * When not in memory, the database will be stored in a platform specific folder
- * - on Android it will use [Context.getDatabaseName](https://developer.android.com/reference/android/content/Context#getDatabasePath(java.lang.String))
+ * - on Android it will use [Context.getCacheDir](https://developer.android.com/reference/android/content/Context#getCacheDir()). It can
+ * also be an absolute path to a file.
  * - on MacOS, it will use "Application Support/databases/name"
  * - on the JVM, it will use "System.getProperty("user.home")/.apollo"
  * - on JS/Wasm, this argument is unused